Add authorization to REST
- Add access mode to the user principal of the authorized request wrapper
- Start transactions with the access mode of this user principal acquired
through the http request / context.
- Add a pass-through authorization filter when auth is disabled that adds
an authorized request wrapper with full access mode.
