+++ community/server/src/test/java/org/neo4j/server/rest/security/AuthenticationDocIT.java
@@ -23,7 +23,6 @@ import com.sun.jersey.core.util.Base64;
 import org.codehaus.jackson.JsonNode;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -250,28 +249,26 @@ public class AuthenticationDocIT extends ExclusiveServerTestBase
         assertThat( firstError.get( "message" ).asText(), equalTo( "Too many failed authentication requests. Please wait 5 seconds and try again." ) );
     }
 
-    // TODO: Enable this test when we have authorization in place
     @Test
-    @Ignore
     public void shouldNotAllowDataAccessForUnauthorizedUser() throws Exception
     {
         // Given
-        startServerWithConfiguredUser(); // TODO: The user for this test should not have read access
+        startServer( true ); // The user should not have read access before changing the password
 
         // When
         HTTP.Response response =
-                HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "secret" ) ).POST(
+                HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "neo4j" ) ).POST(
                         server.baseUri().resolve( "authentication" ).toString(),
-                        HTTP.RawPayload.quotedJson( "{'username':'neo4j', 'password':'secret'}" )
+                        HTTP.RawPayload.quotedJson( "{'username':'neo4j', 'password':'neo4j'}" )
                 );
 
         // When & then
-        assertEquals( 403, HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "secret" ) )
+        assertEquals( 403, HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "neo4j" ) )
                 .POST( server.baseUri().resolve( "db/data/node" ).toString(),
                         RawPayload.quotedJson( "{'name':'jake'}" ) ).status() );
-        assertEquals( 403, HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "secret" ) )
+        assertEquals( 403, HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "neo4j" ) )
                 .GET( server.baseUri().resolve( "db/data/node/1234" ).toString() ).status() );
-        assertEquals( 403, HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "secret" ) )
+        assertEquals( 403, HTTP.withHeaders( HttpHeaders.AUTHORIZATION, challengeResponse( "neo4j", "neo4j" ) )
                 .POST( server.baseUri().resolve( "db/data/transaction/commit" ).toString(),
                         RawPayload.quotedJson( "{'statements':[{'statement':'MATCH (n) RETURN n'}]}" ) ).status() );
     }
